+++ FEV_KEGG/Experiments/48.py
@@ -106,7 +106,7 @@ if __name__ == '__main__':
     output.append('')
     output.append(', '.join(cladeA.ncbiNames) + ':')
     output.append('')
-    cladeAEcGraph = cladeA.coreMetabolism()
+    cladeAEcGraph = cladeA.coreMetabolism(majorityPercentageCoreMetabolism)
     cladeAEcCount = len(cladeAEcGraph.getECs())
     output.append( 'core metabolism ECs: ' + str(cladeAEcCount) )
     output.append('')